fix(web): fix autocomplete prefix behavior
Always try to complete the current token if it is a terminal
(e.g., true, false, unknown, and error after a : or an = sign).
Autocomplete still only starts without being explicitly invoked if there
is a QualifiedName to complete.

Discard completions with a shorter prefix than the current token,
because they would be filtered out by CodeMirror anyways.