feat(web): xtext formatter client
Uses the xtext formatted on the server to format the document.

Also adds the capability to take (delta) changes from the server and
apply them before any pending local changes, then replay the changes.
This means that the server-side formatter is effectively acting as a
second user who is editing the document.