chore: add copyright headers
Make sure we obey the REUSE (https://reuse.software) specification and the
origin, copyright owner, and license of all files are clearly marked.

The whole project is under the EPL-2.0, except for trivial files where copyright
is not applicable that are marked with the CC0-1.0 license. Moreover, code
included from third parties is also available under the respective license.

chore: add CONTRIBUTORS.md

List all authors and supporting organizations in accordance with the REUSE
specification.