chore: clarify permissing licensing
For permissive licenses like MIT and BSD, OR EPL doesn't make sense, because the
copyright original permission notice must be retained at all times.

For code simply copied with minimal modifications that are unrelated to
Refinery, we can just use the plain MIT license or BSD-3-clause license.
For code with complex modifications related to Refinery, we can instead use the
MIT AND EPL-2.0 SPDX expression, which forces downstream users to retain the
original MIT notices as required by the MIT license, but also extends Refinery's
copyleft to our modifications.