fix: VIATRA projection indexer error
When a projection indexer is constructed for a production node, the projection
memory is only populated if changes are being propagated. The cache doesn't get
populated even if changes are flushed afterwards. This not only returns invalid
query results, but also a duplicate deletion exception will be thrown when the
production node tries to delete a tuple from the index memory.

To counteract this issue, we enable update propagation while a matcher (and its
associated indexers) are being created.