feat: predicates as reference types
Now unary predicates may be used as types for cross-references.

This might create circular references between predicate, but no differently than
when the user manually creates predicates with circular references.

The reference types of containment and container references still have to be
classes, because we use the class hierarchy to encode contained types.