fix: upper and lower scopes
Make sure the scope bound guides the best-first search and is always obeyed even
if it can't be fully represented using the scoped partial model abstraction.

For representable bounds (classes and unary predicate that are always fully
knwon to be false or true in the initial model) this leads to duplicated
computations.