refactor(reasoning): candidate view should always be a refinement of partial
Merge the candidate may/must queries with the partial may/must queries so that
reasoning in the candidate model uses "Gentzen-style" assumptions about the
candidate values of partal symbols that would be computed as unknown.

This should not lead to performance degradation, because the corresponding joins
were already being computed in the stopping criterion. In many cases, the extra
clauses are immediately optimized away by the query optimizer anyways.