refactor(reasoning): concretize all existing multi-objects
If we only create concrete objects for object with count=1 (i.e., objects that
are not multi-objects), we may miss predicates with ERROR values on surely
existing but count>1 multi-objects in the concretization view.

Adding a concrete object for every surely existing multi-object still break
refinement from the partial to the candidate view (as object counts will be
off), but at least lets us preserve predicates with ERROR values from the
partial view in the candidate view.

This influences how the objective function is calculated, but doesn't seem to
make the design space explorer get lost in practice.