feat(frontend): use font metrics from browser for visualization
We wrap each text node in a forcedsize="TRUE" table for Graphviz, so we aren't
limited to using the characters which the Graphviz WASM binary knows the font
metrics for.

This allows properly displaying Unicode strings in the visualization in the web
browser and in PNG exports. However, these characters still won't show up in the
PDF export, because we do not embed the corresponding fonts. SVG redering
compatibility varies and may introduce layout error if the renderer uses a
differnt fallback font than the web browser.