fix(generator): InvalidProblemException resource URI
Make sure we set the correct resource URI for InvalidProblemException even if
the validation issues come from a TracedException, so that all reported file
offsets are actually relative to the Problem resoruce.

If we report offsets from other resources, CodeMirror may throw and exception as
we try to place an out-of-range diagnostic for the editor in the frontend.