refactor(language): nodes should not clash with variables
If a user creates a node with the same name as a variable, still consider the
referece in a predicate to refer to the variable, as non-atom nodes can't be
referenced in predicates. However, if a user creates an atom with the same name
as variable, consider the reference in a predicate to refer to the atom.